requirements.txt
# Display driver for ST7789 TFT
ST7789==0.0.19

# Image processing
Pillow>=10.0.0

# GPIO control (Raspberry Pi)
RPi.GPIO>=0.7.0
gpiozero>=2.0.0

# Camera
picamera2>=0.3.0

# QR code generation and scanning
qrcode[pil]>=7.4.0
pyzbar>=0.1.9

# Bitcoin operations
bitcoinlib>=0.6.0
ecdsa>=0.18.0

# SPI interface
spidev>=3.6

# Environment configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional, for development)
mypy>=1.5.0

# Additional utilities
numpy>=1.24.0
